# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IronLog fitness tracking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface AuthCredentials {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
}

export interface WorkoutSet {
  id: number;
  workoutId: number;
  exerciseName: string;
  weight: number;
  reps: number;
  rpe?: number | null;
}

export interface CreateSetBody {
  exerciseName: string;
  weight: number;
  reps: number;
  rpe?: number | null;
}

export interface CreateWorkoutBody {
  date: string;
  notes?: string | null;
  sets: CreateSetBody[];
}

export interface WorkoutWithSets {
  id: number;
  userId: number;
  date: string;
  notes?: string | null;
  sets: WorkoutSet[];
  totalVolume: number;
  exerciseCount: number;
}

export interface DashboardStats {
  weeklyWorkoutCount: number;
  totalWorkouts: number;
  recentWorkouts: WorkoutWithSets[];
  weeklyVolume: number;
  allTimeVolume: number;
}

export interface Quote {
  text: string;
  author: string;
}

export type ListWorkoutsParams = {
  limit?: number;
  offset?: number;
};