# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IronLog fitness tracking API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a new account
 */
export const signupBodyPasswordMin = 6;

export const SignupBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(signupBodyPasswordMin),
});

/**
 * @summary Login to an existing account
 */
export const loginBodyPasswordMin = 6;

export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(loginBodyPasswordMin),
});

export const LoginResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
});

/**
 * @summary Get current authenticated user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
});

/**
 * @summary List workouts for the authenticated user
 */
export const listWorkoutsQueryLimitDefault = 20;
export const listWorkoutsQueryOffsetDefault = 0;

export const ListWorkoutsQueryParams = zod.object({
  limit: zod.coerce.number().default(listWorkoutsQueryLimitDefault),
  offset: zod.coerce.number().default(listWorkoutsQueryOffsetDefault),
});

export const ListWorkoutsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  date: zod.string(),
  notes: zod.string().nullish(),
  sets: zod.array(
    zod.object({
      id: zod.number(),
      workoutId: zod.number(),
      exerciseName: zod.string(),
      weight: zod.number(),
      reps: zod.number(),
      rpe: zod.number().nullish(),
    }),
  ),
  totalVolume: zod.number(),
  exerciseCount: zod.number(),
});
export const ListWorkoutsResponse = zod.array(ListWorkoutsResponseItem);

/**
 * @summary Create a new workout
 */
export const CreateWorkoutBody = zod.object({
  date: zod.coerce.date(),
  notes: zod.string().nullish(),
  sets: zod.array(
    zod.object({
      exerciseName: zod.string(),
      weight: zod.number(),
      reps: zod.number(),
      rpe: zod.number().nullish(),
    }),
  ),
});

/**
 * @summary Get a single workout
 */
export const GetWorkoutParams = zod.object({
  id: zod.coerce.number(),
});

export const GetWorkoutResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  date: zod.string(),
  notes: zod.string().nullish(),
  sets: zod.array(
    zod.object({
      id: zod.number(),
      workoutId: zod.number(),
      exerciseName: zod.string(),
      weight: zod.number(),
      reps: zod.number(),
      rpe: zod.number().nullish(),
    }),
  ),
  totalVolume: zod.number(),
  exerciseCount: zod.number(),
});

/**
 * @summary Delete a workout
 */
export const DeleteWorkoutParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardStatsResponse = zod.object({
  weeklyWorkoutCount: zod.number(),
  totalWorkouts: zod.number(),
  recentWorkouts: zod.array(
    zod.object({
      id: zod.number(),
      userId: zod.number(),
      date: zod.string(),
      notes: zod.string().nullish(),
      sets: zod.array(
        zod.object({
          id: zod.number(),
          workoutId: zod.number(),
          exerciseName: zod.string(),
          weight: zod.number(),
          reps: zod.number(),
          rpe: zod.number().nullish(),
        }),
      ),
      totalVolume: zod.number(),
      exerciseCount: zod.number(),
    }),
  ),
  weeklyVolume: zod.number(),
  allTimeVolume: zod.number(),
});

/**
 * @summary Get a random motivational quote
 */
export const GetMotivationalQuoteResponse = zod.object({
  text: zod.string(),
  author: zod.string(),
});