# lib/api-zod/src/generated/types/createSetBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IronLog fitness tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSetBody {
  exerciseName: string;
  weight: number;
  reps: number;
  rpe?: number | null;
}