# lib/api-zod/src/generated/types/createWorkoutBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IronLog fitness tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateSetBody } from "./createSetBody";

export interface CreateWorkoutBody {
  date: Date;
  notes?: string | null;
  sets: CreateSetBody[];
}