# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IronLog fitness tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutWithSets } from "./workoutWithSets";

export interface DashboardStats {
  weeklyWorkoutCount: number;
  totalWorkouts: number;
  recentWorkouts: WorkoutWithSets[];
  weeklyVolume: number;
  allTimeVolume: number;
}