# lib/api-zod/src/generated/types/workoutWithSets.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IronLog fitness tracking API
 * OpenAPI spec version: 0.1.0
 */
import type { WorkoutSet } from "./workoutSet";

export interface WorkoutWithSets {
  id: number;
  userId: number;
  date: string;
  notes?: string | null;
  sets: WorkoutSet[];
  totalVolume: number;
  exerciseCount: number;
}